requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
dataclasses==0.7
deprecation==2.0.7
eyeD3==0.9.5
filetype==1.0.6
Flask==1.1.2
get==2019.4.13
google-api-core==1.16.0
google-api-python-client==1.8.3
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
youtube-dl==2020.5.8
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
deprecation==2.1.0
eyeD3==0.9.5
filetype==1.0.7
Flask==1.1.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
youtube-dl
